feat: introduce topos-node crate
