+++ crates/topos/src/components/node/services.rs
@@ -1,2 +1,1 @@
-pub(crate) mod process;
 pub(crate) mod status;
